Move getGatorByName, getFaculty and getFaculties to goldleRunCrew, to remove unncessary functions in main Goldle class
